[TASK] - better fix for prechecks
